Create and add JokesLib + Update some dependencies
Used Android studio to create Java library JokesLib. This library is responsible to provide the jokes.
So a class Jokes was created along with a list of jokes taken from a webpage and a method that uses Random to
extract a random joke from that list.

on app build.gradle implementation project(":jokeslib"). On MainActivity an object Jokes was created to be able
to call the method getJoke().

Dependencies such junit and play-services-ad were updated.

Note: gradle issue a warning: warning: [options] bootstrap class path not set in conjunction with -source 1.7
1 warning." Needs to be checked